+++ package.json
@@ -1,11 +1,10 @@
 {
   "dependencies": {
-    "@discordjs/opus": "^0.8.0",
-    "@discordjs/voice": "^0.16.1",
-    "@distube/ytdl-core": "^4.13.3",
+    "@discordjs/opus": "^0.9.0",
+    "@discordjs/voice": "^0.17.0",
     "discord-hybrid-sharding": "^2.1.0",
     "discord-ytdl-core": "^5.0.4",
-    "discord.js": "^14.14.1",
+    "discord.js": "^14.15.3",
     "express": "^4.18.2",
     "ffmpeg-static": "^5.1.0",
     "fluent-ffmpeg": "^2.1.2",
@@ -15,9 +14,9 @@
     "music-metadata": "^7.13.1",
     "node-gyp": "^9.3.1",
     "pg": "^8.8.0",
+    "play-audio": "^0.5.2",
     "winston": "^3.8.2",
     "winston-daily-rotate-file": "^4.7.1",
-    "ytdl-core": "^4.11.3",
-	"play-audio": "^0.5.2"
+    "ytdl-core": "^4.11.5"
   }
 }
